Homer Simpson
hello! My name is Herick, I'm here presenting my first drawing (project) that I developed more for learning. I'm a Software engineering student and to practice I decided to develop a design using HTML and CSS, that's when I decided to draw Homer Simpson. And that was the end result, I hope it turned out cool.